Change YouTube content end-point
