+++ package.json
@@ -27,6 +27,7 @@
     "axios": "^1.5.0",
     "cheerio": "1.0.0-rc.12",
     "date-fns": "2.29.3",
+    "front-matter": "^4.0.2",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "masonic": "^3.7.0",
