+++ utils/sheets.ts
@@ -1,18 +1,28 @@
-import { notion } from './notion';
+import axios from 'axios';
+
+interface RowData {
+  idx: string;
+  video_id: string;
+  description: string;
+  comment: string;
+  created: string;
+  title: string;
+}
 
 export async function getSheetData(start?: number, count?: number) {
-  const response = await notion.databases.query({ database_id: process.env.NOTION_DATABASE_ID_YOUTUBE! });
-  const rowsData = response.results.map((result: any) => {
-    return {
-      idx: result.properties.idx?.formula?.string || '',
-      video_id: result.properties.video_id?.rich_text[0]?.plain_text || '',
-      summary: result.properties.summary?.rich_text[0]?.plain_text || '',
-      blockquote: result.properties.blockquote?.rich_text[0]?.plain_text || '',
-      created: result.properties.created?.date?.start || '',
-      subject: result.properties.subject?.title[0]?.plain_text || '',
-    };
-  });
-  const sortedRowsData = rowsData.sort((a, b) => b.idx.localeCompare(a.idx));
+  const response = await axios.get(`${process.env.NEXT_PUBLIC_API_URL}/api/youtube`);
+  const filesData = response.data;
+
+  const rowsData: RowData[] = filesData.map((data: any) => ({
+    idx: data.attributes.idx,
+    video_id: data.attributes.video_id,
+    description: data.attributes.description,
+    comment: data.attributes.comment,
+    created: data.attributes.created,
+    title: data.attributes.title,
+  }));
+
+  const sortedRowsData = rowsData.sort((a: RowData, b: RowData) => b.idx.localeCompare(a.idx));
 
   if (start !== undefined && count !== undefined) {
     return sortedRowsData.slice(start, start + count);
